Migrado preguntar a su propio componente. Ahora solo renderiza preguntar a usuarios logueados.
